requirements.txt
torch>=2.1
transformers==4.47.0
accelerate==1.3.0
peft==0.13.2
huggingface_hub>=0.23.0
pillow>=10.0.0
scipy==1.15.1
